Watch fixed and improved

+++ src/node/lib/watch.js
@@ -4,14 +4,20 @@ var commons = require("@vimlet/commons");
 var meta = require("../index.js");
 var fs = require("fs-extra");
 
-exports.watch = function(include, exclude, output, data, callback) {
-  var watcher = watch(include);
+exports.watch = function(include, exclude, output, data) {
+  console.log("Watching:",include);
+  var watcher = watch(include,{
+    events: ['add', 'change', 'unlink', 'addDir', 'unlinkDir']
+  });
   watcher.on('change', function(filePath, stat) {
     if (!isExcluded(exclude, filePath)) {
       // Relative output is where the template will be saved after parsed
       var relativeOutput = getRelativeOutput(include, output, filePath);
       // Parse modified file
-      meta.parseTemplateWrite(null, filePath, null, relativeOutput, data, false, callback);
+      meta.parseTemplateWrite(null, filePath, null, relativeOutput, data, false, function(){
+        console.log("File modified:");
+        console.log(filePath + " => " + path.join(relativeOutput,path.basename(filePath, path.extname(filePath))));
+      });
     }
   });
   watcher.on('add', function(filePath, stat) {
@@ -19,17 +25,47 @@ exports.watch = function(include, exclude, output, data, callback) {
       // Relative output is where the template will be saved after parsed
       var relativeOutput = getRelativeOutput(include, output, filePath);
       // Parse modified file
-      meta.parseTemplateWrite(null, filePath, null, relativeOutput, data, false, callback);
+      meta.parseTemplateWrite(null, filePath, null, relativeOutput, data, false, function(){
+        console.log("File added:");
+        console.log(filePath + " => " + path.join(relativeOutput, path.basename(filePath, path.extname(filePath))));
+      });
     }
   });
   watcher.on('unlink', function(filePath, stat) {
     if (!isExcluded(exclude, filePath)) {
       // Relative output is where the template will be saved after parsed
       var relativeOutput = getRelativeOutput(include, output, filePath, true);
-      console.log("filePath",filePath);
-      console.log("relativeOutput",relativeOutput);
       var parsedPath = path.join(relativeOutput, path.basename(filePath,path.extname(filePath)));
-      fs.unlinkSync(parsedPath);
+      if(fs.existsSync(parsedPath)){
+        fs.unlinkSync(parsedPath);
+        console.log("File deleted:");
+        console.log(parsedPath);
+      }
+    }
+  });
+  watcher.on('addDir', function(filePath, stat) {
+    var relativeOutput = getRelativeOutput(include, output, filePath);
+    fs.mkdirs(path.join(relativeOutput, path.basename(filePath)), function(){
+      console.log("Folder created:");
+      console.log(filePath, "=>", path.join(relativeOutput, path.basename(filePath)));
+    });
+  });
+  watcher.on('unlinkDir', function(filePath, stat) {
+    var relativeOutput = getRelativeOutput(include, output, filePath, true);
+    fs.remove(path.join(relativeOutput, path.basename(filePath)), function(){
+      console.log("Folder removed:");
+      console.log(path.join(relativeOutput, path.basename(filePath)));
+    });
+  });
+  watcher.on('error', function(error) {
+    if (process.platform === 'win32' && error.code === 'EPERM') {
+      // fs.open(path, 'r', function(err, fd) {
+      //   if (fd) fs.close(fd);
+      //   if (!err) broadcastErr(error);
+      // });
+      console.log("ERROR","Deleting an empty folder doesn't fire on windows");
+    } else {
+      broadcastErr(error);
     }
   });
 };
@@ -37,10 +73,10 @@ exports.watch = function(include, exclude, output, data, callback) {
 /*
 @function getRelativeOutput [Get path relative to output]
 @param include [Include patterns]
- @param output
-  @param filePath
-  @param deleted [Flag to know if the file was deleted so it skips file in patter check]
- */
+@param output
+@param filePath
+@param deleted [Flag to know if the file was deleted so it skips files in pattern check]
+*/
 function getRelativeOutput(include, output, filePath, deleted) {
   var relativeOutput;
   if (!Array.isArray(include)) {
@@ -52,7 +88,7 @@ function getRelativeOutput(include, output, filePath, deleted) {
     }
   } else {
     include.forEach(function(incl) {
-      if (commons.io.isInPattern(filePath, incl)) {
+      if (commons.io.isInPattern(filePath, incl) || deleted) {
         var rootFromPattern = commons.io.getRootFromPattern(incl);
         // Relative output is where the template will be saved after parse
         relativeOutput = path.dirname(path.relative(rootFromPattern, filePath));
@@ -68,7 +104,7 @@ function getRelativeOutput(include, output, filePath, deleted) {
 @param excluded [exclude patterns]
 @param filePath
 @return boolean
- */
+*/
 function isExcluded(excluded, filePath) {
   if (!Array.isArray(excluded)) {
     return commons.io.isInPattern(filePath, excluded);
